+++ src/SeablastMysqli.php
@@ -66,8 +66,14 @@ class SeablastMysqli extends mysqli
         $this->logPath = Debugger::$logDirectory . '/query_' . date('Y-m') . '.log';
     }
 
-    /* public for SeablastMysqliStmt */
-    // param \mysqli|\mysqli_stmt $dbCall
+    /**
+     * This method is public, so that SeablastMysqliStmt may use it.
+     *
+     * @param bool $result
+     * @param string $trimmedQuery
+     * @param \mysqli|\mysqli_stmt $dbCall
+     * @return void
+     */
     public function addStatement(bool $result, string $trimmedQuery, object $dbCall): void
     {
         if ($result !== false) {
@@ -75,24 +81,21 @@ class SeablastMysqli extends mysqli
                 $trimmedQuery;
             return;
         }
-      //  $this->statementList[] = ($result === false ? 'failure => ' : '') . $trimmedQuery;
-       //     if ($result === false) {
-                $this->databaseError = true;
-                $dbError = ['query' => $trimmedQuery, 'Err#' => $dbCall->errno, 'Error:' => $dbCall->error];
-                Debugger::barDump($dbError, 'Database error', [Dumper::TRUNCATE => 1500]); // longer than 150 chars
-                Debugger::log('Database error' . print_r($dbError, true), ILogger::ERROR);
-                $this->statementList[] = "failure {$dbCall->errno}: {$dbCall->error} => " .$trimmedQuery;
-                $this->logQuery("{$trimmedQuery} -- {$dbCall->errno}: {$dbCall->error}");
-        //    }
+        $this->databaseError = true;
+        $dbError = ['query' => $trimmedQuery, 'Err#' => $dbCall->errno, 'Error:' => $dbCall->error];
+        Debugger::barDump($dbError, 'Database error', [Dumper::TRUNCATE => 1500]); // longer than 150 chars
+        Debugger::log('Database error' . print_r($dbError, true), ILogger::ERROR);
+        $this->statementList[] = "failure {$dbCall->errno}: {$dbCall->error} => " . $trimmedQuery;
+        $this->logQuery("{$trimmedQuery} -- {$dbCall->errno}: {$dbCall->error}");
     }
-    
+
     /**
      * Override the prepare method to return LoggedMysqliStmt.
      *
      * @param string $query
      * @return SeablastMysqliStmt|false
      */
-    public function prepare(string $query)
+    public function prepare($query)
     {
         // Call parent method to prepare the statement
         $stmt = parent::prepare($query);
@@ -123,16 +126,6 @@ class SeablastMysqli extends mysqli
         }
         try {
             $result = parent::query($trimmedQuery, $resultmode);
-            /*
-            $this->statementList[] = ($result === false ? 'failure => ' : '') . $trimmedQuery;
-            if ($result === false) {
-                $this->databaseError = true;
-                $dbError = ['query' => $trimmedQuery, 'Err#' => $this->errno, 'Error:' => $this->error];
-                Debugger::barDump($dbError, 'Database error', [Dumper::TRUNCATE => 1500]); // longer than 150 chars
-                Debugger::log('Database error' . print_r($dbError, true), ILogger::ERROR);
-                $this->statementList[] = "{$this->errno}: {$this->error}";
-                $this->logQuery("{$trimmedQuery} -- {$this->errno}: {$this->error}");
-            } */
             $this->addStatement ((bool) $result, $trimmedQuery, $this);
             return $result;
         } catch (\mysqli_sql_exception $e) {