+++ src/SeablastView.php
@@ -165,17 +165,21 @@ class SeablastView
             && (class_exists($this->model->getConfiguration()->getString(SeablastConstant::TRANSLATE_CLASS)))
         ) {
             $translatorClass = $this->model->getConfiguration()->getString(SeablastConstant::TRANSLATE_CLASS);
+            Debugger::barDump($translatorClass,'Translate class');
             $translator = new $translatorClass($this->model->getConfiguration());
 
             //if (!method_exists($latte, 'addExtension')) {
             // for Latte 2 and 3
             if (method_exists($translator, 'translate')) {
+                Debugger::barDump('Translator exists');
                 //$latte->addFunction('translate', function ($text) use ($translator) {
                 //    return $translator->translate($text);
                 //});
+                Debugger::barDump($latte, 'before filter');
                 $latte->addFilter('translate', function (string $s) use ($translator) {
                     return $translator->translate($s);
                 });
+                Debugger::barDump($latte, 'after filter');
             }
             /*} else {
                 // for Latte 3, i.e. PHP/8.0-8.4
@@ -186,6 +190,8 @@ class SeablastView
                 );
                 $latte->addExtension($extension);
             }*/
+        } else {
+            Debugger::barDump('NO TRANSLATION');
         }
 
         // Maybe only for PHP8+