+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - TableViewModel for admin.latte
 - SeablastMysqli->prepare() is logged the same way as SeablastMysqli->query()
-- TODO figure out - SeablastMysqli->prepare()->execute()
+- Todo figure out - SeablastMysqli->prepare()->execute()
 - Run super-linter and composer-dependencies workflows at 6:30 AM UTC on the 15th of every month
 
 ### `Changed` for changes in existing functionality