+++ CHANGELOG.md
@@ -10,9 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `Added` for new features
 
 - TableViewModel for admin.latte
-- SeablastMysqli->prepare() is logged the same way as SeablastMysqli->query()
-- Todo figure out - SeablastMysqli->prepare()->execute()
+- SeablastMysqli->prepare() is logged the same way as SeablastMysqli->query() not SeablastMysqli->prepare()->execute() however
 - Run super-linter and composer-dependencies workflows at 6:30 AM UTC on the 15th of every month
+- Make sure that the database Tracy BarPanel is displayed when DbmsException is thrown
 
 ### `Changed` for changes in existing functionality