+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Security` in case of vulnerabilities
 
-## [0.2.7] - 2025-01-25
+## [0.2.7] - 2025-01-26
 
 SeablastMysqli->prepare() is logged, the database Tracy BarPanel is displayed when DbmsException is thrown.
 
@@ -30,6 +30,7 @@ SeablastMysqli->prepare() is logged, the database Tracy BarPanel is displayed wh
 - SeablastMysqli->prepare() is logged the same way as SeablastMysqli->query() (not SeablastMysqli->prepare()->execute() however)
 - Run super-linter and composer-dependencies workflows at 6:30 AM UTC on the 15th of every month
 - Make sure that the database Tracy BarPanel is displayed when DbmsException is thrown
+- SeablastPdo added to be logged the same way as SeablastMysqli->prepare()
 
 ### Changed